requirements.txt
Flask
flask-cors
pydub
googletrans==4.0.0-rc1
openai-whisper
yt-dlp
gunicorn
